Remove -[autorelease] from OFBlock.
The reason is that a block might be an NSBlock on OS X where it would
add the block to an NSAutoreleasePool. This change forces the developer
to use [OFAutoreleasePool addObject: block], which will also work if
the block is an NSBlock. Thus, this change is to force developers to
write portable code.