Increase blocks compatibility by replacing NSAutoreleasePool.
NSAutoreleasePool is replaced by OFAutoreleasePool now so that there is
no conflict because OFBlocks are used (blocks are OFBlocks as soon as
ObjFW is linked). An application expecting an NSBlock, but getting an
OFBlock because ObjFW is linked, would fail to autorelease the block
otherwise, as the block would be in an OFAutoreleasePool. By replacing
NSAutoreleasePool with OFAutoreleasePool, the application will still
properly free the autoreleased block.

This will also make it easier to write applications using both,
Foundation and ObjFW, since the developer does not need to take care
about two different autorelease pools, but they are united instead.