Fix two bugs in sel_registerName.
The first was that sel_registerName would crash if no selector has ever
been registered before. This caused a crash on NetBSD in has_load, as
the first class being loaded there (Protocol) had no selectors and
has_load uses sel_registerName in order to get the load selector.

The second one was that the hashtable was read without a lock to see if
the selector has already been registered before.