OFRunLoop: Remove the key for the stream.
As the stream is used as a key, this references the stream. This
resulted in the stream never being deallocated, even though it was not
used anymore. The advantage of not removing it was that the next async
read on the stream would have been slightly faster by saving one malloc.