Revert several OFHashMap related commits.
This reverts:
 * "Add a per-hashtable seed."
   (cd49c7fcb4b8dca62dccdbf6545d37d86b639eef)
 * "OFMapTable: Reseed on resize."
   (5b545165691cd9ad838c9125b02ff7c857bd4b99)
 * "Add of_random()."
   (9810f191b4e8c1612e7e55095c87e589660fbdba)
 * "OFMapTable: Rotate hash by a random number of bits"
   (3bef412217232d01f202794f7175c130afa1df8f)

It turned out that these were slowing down all classes based on
OFHashMap significantly. As there is still no proof that the seeded
one-at-a-time hash is vulnerable to DoS, these changes were overly
paranoid. While overly paranoid is usually ok, it is not if it severely
impacts performance.