OFNumber: Remove methods for arithmetic
These methods encouraged doing arithmetic on objects rather than getting
the C type value, doing the arithmetic and creating a new object. They
didn't do anything else internally, and so doing this manually isn't
slower - it's even faster if you do multiple operations, as no temporary
objects are created then.

Another advantage of this is that there are no longer uncertainties
about types of different OFNumber objects when doing arithmetic, as the
user has to be explicit about the types when getting the C type value
and thus knows which types will be used for the arithmetic.