README-WINDOWS.md: Complete rewrite for MSYS2
MSYS2 is currently the best way to use MinGW on Windows and also the
best way to keep it up to date. It offers Win64 support as well and can
be used to install a lot of useful tools, for example Git, and usually
has quite recent versions of all packages.