OFBlock: Fix two bugs
The first was that _Block_object_{assign,dispose} would only be a NOP
when called from a byref helper instead of only being a NOP when called
for an object from a byref helper.

The second was that byref helpers would even be called when there are no
byref helpers (bit 25 in the byref flags needs to be checked -
conveniently, bit 25 indicates copy/dispose helpers in the block literal
flags, too, so that define can be reused).