typedef socklen_t to int on MorphOS
The previous approach of including <pthread.h> - which has the typedef -
turned out to be wrong, as that header is only a stub and including it
breaks things.

Since all the functions taking a socklen_t on other platforms take an
int on MorphOS, creating a typedef to int is the proper way.