Work around OpenBSD 6.2 not linking the unwinder
OpenBSD 6.2 switched to Clang as the primary compiler on i386 and amd64.
This switch also means that their clang no longer links in libgcc,
meaning it links in no unwinder. The unwinder is now in libc++abi, which
we now need to link in manually, alongside with pthreads, as libc++abi
depends on pthreads.