Remove Xcode project
It is too annoying to maintain two build systems and keep them in sync
with all the flags, especially as Xcode sometimes gives the same flags
different names.

An Xcode project to build and run the tests for iOS will be added back
later: This will depend on the .framework and all tests as an .a being
built by the make-based build system and then just create an app from
that.