OFLHAArchive: Prefill the sliding window with ' '
Apparently, there are two incompatible variants of LHA: One where the
sliding window is prefilled with zeros and one where it is prefilled
with spaces. It seems prefilling it with spaces is the original way,
while LHA on Amiga prefilled it with zeros.

As it is impossible to detect which one was used, a flag will be added
later to provide an Amiga compatibility mode.