OFThread: Allow specifying a name before starting
This allows specifying a name before the thread gets started, so that
the name can be decided by whoever starts the thread, rather than just
by the thread itself once it's running.

This is especially useful as some operating systems do not support
changing the name of the thread once it's running.