runtime: Increase the library major version
Now that the runtime structs are opaque, apps should stop using them
directly. Technically, making them opaque does not break the ABI, but as
soon as the structs change, the ABI will break. And the entire point of
making them opaque is that they can change, hence increase the library
major version now.