Remove OFIPStreamSocket
This was a bad idea, as while SCTP can be used via the byte
stream-oriented API, it will still work in messages. But OFStream is not
afraid of breaking up and reassembling reads, as it is meant for byte
streams. Therefore, implementing SCTP as a subclass of OFStream would
degrate it to a stream-oriented socket instead of a message-oriented
socket, rendering one of the main features of SCTP void.