Make of_random() a function
When arc4random() is unavailable, either random() or rand() is used and
both need to be seeded. If of_random() is a macro, it needs to be
(re)seeded every time, as it's unknown whether it has already been
seeded. As it is seeded with gettimeofday() due to the lack of a better
initial seed, this means every call returns the first state for the
current time, which is very predictable. random() and rand() are both
not cryptographic, but this should at least make it a little bit better
now.

FossilOrigin-Name: c1fe4b2b772d5ccfda3f9c825cdf694ebcab30958c47226f7949daa1f71981b9