runtime: Compile a copy of {mutex,once,tlskey}.m
This makes sure they are compiled with the flags that are required for
objfwrt(68k).library, while being compiled with the regular flags for
ObjFW itself.

This does not cause problems for static libraries or shared libraries,
as there the same flags are used for the runtime and ObjFW, resulting in
in two identical .o files, just as before.

FossilOrigin-Name: c1354656bcd7f892ed9ff11c6cff97436eb823de160a3260a1432fc3f0bdc053